SuperArchitect/requirements.txt
requests
anthropic
openai
google-generativeai
aiohttp
asyncio
python-dotenv
PyYAML